% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data()
}
\value{
a list of two tibbles: df_main and df_details
}
\description{
Function to get the necessary data from the OGD portal.
}
\examples{
data <- get_data()
}
