% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reactables_candidates.R
\name{get_reactable_candidates}
\alias{get_reactable_candidates}
\title{get_reactable_candidates}
\usage{
get_reactable_candidates(df)
}
\arguments{
\item{df}{filtered data to be shown}
}
\value{
a reactable
}
\description{
function to generate the main table with a list of candidates
}
