% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reactables_candidates.R
\name{get_reactable_details}
\alias{get_reactable_details}
\title{get_reactable_details}
\usage{
get_reactable_details(candidate)
}
\arguments{
\item{candidate}{}
}
\value{
a reactable
}
\description{
function to generate the reactable for the details about the candidate
}
