% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KALI-package.R
\docType{package}
\name{KALI-package}
\alias{KALI}
\alias{KALI-package}
\title{KALI: Shiny App zu Kandidierenden des Zürcher Gemeinderates}
\description{
Golem-paketierte Shiny App.
}
\author{
\strong{Maintainer}: AUTHOR_FIRST AUTHOR_LAST \email{AUTHOR@MAIL.COM}

}
\keyword{internal}
