% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_update_chart.R
\name{update_chart}
\alias{update_chart}
\title{update_chart}
\usage{
update_chart(data, type, session)
}
\arguments{
\item{data}{data as tibble to be sent}

\item{type}{string, specifies which message handler/chart receives the data}

\item{session}{}
}
\value{
na
}
\description{
sends data as json as a custom message, which can then be
received by JS/the sszvis chart; needs an appropriate message handler
function specific to each chart
}
